DevOps/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
docker==6.1.3
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
protobuf==4.21.12
python-decouple==3.8
requests==2.31.0
secure-smtplib==0.1.1
smmap==5.0.0
urllib3==2.0.4
websocket-client==1.6.1
Werkzeug==2.3.6
zipp==3.16.2
